Moving FeatureExtractionPipeline into its own file.
